refactor: ajust directory and files
