--- libs/feature/src/lib/components/form-auth-card/index.ts
@@ -1,1 +1,0 @@
-export * from './form-auth-card';
