--- libs/feature/src/lib/components/form-button/index.ts
@@ -1,1 +1,0 @@
-export * from './form-button.component';
