--- libs/feature/src/lib/components/form-create-auth/index.ts
@@ -1,1 +1,0 @@
-export * from './form-create-auth';
