--- libs/feature/src/lib/components/form-create-user/index.ts
@@ -1,1 +1,0 @@
-export * from './form-create-user';
