+++ libs/feature/src/lib/components/index.ts
@@ -1,8 +1,5 @@
 export * from './alert';
 export * from './drawer';
-export * from './form-auth-card';
-export * from './form-button';
-export * from './form-create-auth';
-export * from './form-create-user';
+export * from './form';
 export * from './notification-card';
 export * from './stepper';
